+++ Anoroc-User-Management/Models/Cluster.cs
@@ -11,7 +11,7 @@ namespace Anoroc_User_Management.Services
     
     public class Cluster : DbContext
     {
-        public int Cluster_ID { get; }
+        public long Cluster_ID { get; }
         public List<Location> Coordinates { get; set; }
         public Location Center_Location { get; set; }
         public int Carrier_Data_Points;
@@ -24,7 +24,7 @@ namespace Anoroc_User_Management.Services
             // Create a function that scans through the list of clusters and removes the ones that have been there the longest
         }
 
-        public Cluster(Location loc, int cluster_id)
+        public Cluster(Location loc, long cluster_id)
         {
 
             Coordinates = new List<Location>();
